+++ static/.env
@@ -1,0 +1,1 @@
+REACT_APP_BACKEND="http://localhost:8000"
