+++ static/.gitignore
