--- static/bifrost-front/src/Services/auth.js
@@ -1,6 +1,0 @@
-/***
- * Check if the user is logged in.
- */
-const isLoggedIn = () => {
-
-};
