--- static/js/home.test.js
@@ -1,7 +1,0 @@
-import {react} from 'React';
-import 'home'
-
-
-test('adds 1 + 2 to equal 3', () => {
-  expect(Home).toBe(<div>Hello!</div>)
-});
