+++ static/public/css/sass/home.scss
