+++ static/public/manifest.json
