+++ static/src/Services/auth.js
@@ -1,0 +1,9 @@
+
+/***
+ * Check if the user is logged in.
+ */
+function isLoggedIn() {
+    return false;
+}
+
+export default isLoggedIn
