+++ static/src/index.js
